adjusts the texture viewer proper
